DFSmethods. Pre-repease refactoring.
Release notes
1.Simplest working algorithm for pizza slicing.
