+++ src/main/java/com/google/hashcode/utils/FilesPaths.java
@@ -1,0 +1,22 @@
+package com.google.hashcode.utils;
+
+/**
+ * Contains common files to perform input-output
+ *
+ * @author Grigoriy Lyashenko (Grog).
+ */
+public abstract class FilesPaths {
+    public static final String EXAMPLE_INPUT_FILE_PATH = "inputDataSets/example.in";
+    public static final String SMALL_INPUT_FILE_PATH = "inputDataSets/small.in";
+    public static final String MEDIUM_INPUT_FILE_PATH = "inputDataSets/medium.in";
+    public static final String BIG_INPUT_FILE_PATH = "inputDataSets/big.in";
+
+
+    public static final String OUTPUT_DATA_SET_EXAMPLE_TXT = "outputDataSet/example.txt";
+    public static final String OUTPUT_DATA_SET_SMALL_TXT = "outputDataSet/small.txt";
+    public static final String OUTPUT_DATA_SET_BIG_TXT = "outputDataSet/big.txt";
+    public static final String OUTPUT_DATA_SET_MEDIUM_TXT = "outputDataSet/medium.txt";
+
+    private FilesPaths() {
+    }
+}
