--- src/main/java/com/google/hashcode/utils/InputFiles.java
@@ -1,14 +1,0 @@
-package com.google.hashcode.utils;
-
-/**
- * @author Grigoriy Lyashenko (Grog).
- */
-public abstract class InputFiles {
-    public static final String EXAMPLE_INPUT_FILE_PATH = "inputDataSets/example.in";
-    public static final String SMALL_INPUT_FILE_PATH = "inputDataSets/small.in";
-    public static final String MEDIUM_INPUT_FILE_PATH = "inputDataSets/medium.in";
-    public static final String BIG_INPUT_FILE_PATH = "inputDataSets/big.in";
-
-    private InputFiles() {
-    }
-}
