+++ src/test/java/com/google/hashcode/utils/DFSMethodsTest.java
@@ -10,7 +10,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
-import static com.google.hashcode.utils.InputFiles.EXAMPLE_INPUT_FILE_PATH;
+import static com.google.hashcode.utils.FilesPaths.EXAMPLE_INPUT_FILE_PATH;
 import static org.junit.Assert.assertEquals;
 
 /**
